Fiasco.OC: fix DMA l2-cache coherency problems
By now, Fiasco.OC doesn't make use of the outer L2-cache (PL310) of the
Pandaboard. Thereby we get problems when using DMA memory regions that
are in the cache. This fix uses the right syscall in core's ram-session,
and adds a patch to the Fiasco.OC kernel to use the L2 cache properly.